message sending

+++ MaelstormAPI/Services/Abstractions/IDialogService.cs
@@ -7,6 +7,6 @@ namespace MaelstormApi.Services.Abstractions
     public interface IDialogService
     {
         public Task<List<Dialog>> GetDialogsAsync(int offset = 0, int count = 10);
-        public Task<Dialog> GetDialogAsync(int interlocutorId);
+        public Task<Dialog> GetDialogAsync(long interlocutorId);
     }
 }
